+++ lib/PointyCastleN/src/registry/registration.dart
@@ -49,6 +49,8 @@ import '../../stream/salsa20.dart';
 import '../../stream/salsa2012.dart';
 import '../../stream/salsa208.dart';
 import '../../stream/sic.dart';
+import '../../stream/isaac.dart';
+import '../../stream/grain128.dart';
 
 void registerFactories(FactoryRegistry registry) {
   _registerAsymmetricCiphers(registry);
@@ -145,4 +147,6 @@ void _registerStreamCiphers(FactoryRegistry registry) {
   registry.register(Salsa12Engine.FACTORY_CONFIG);
   registry.register(SICStreamCipher.FACTORY_CONFIG);
   registry.register(HC256Engine.FACTORY_CONFIG);
+  registry.register(Grain128Engine.FACTORY_CONFIG);
+  registry.register(ISAACEngine.FACTORY_CONFIG);
 }